Sentiment analyzer/requirements.txt
Flask
Flask-Cors
pandas
chromadb
matplotlib
Markdown
beautifulsoup4
python-docx
ollama